refactor: move verifier to a separate module
